enhance(frontend): tweak control panel
